+++ src/custom/global.ts
@@ -1,8 +1,9 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import knex from 'knex'
+import { Firestore } from '@google-cloud/firestore'
 
 interface CustomNodeJSGlobal extends NodeJS.Global {
-  firestoreDatabase: FirebaseFirestore.Firestore,
+  firestoreDatabase: Firestore,
   mysqlDatabase    : knex<any, unknown[]>
 }