+++ src/routes/index.ts
@@ -1,0 +1,15 @@
+import { Home } from './home'
+import { Jobs } from './jobs'
+import { JobOffers } from './jobOffers'
+import { Projects } from './projects'
+import { Tasks } from './tasks'
+import { Users } from './users'
+
+export {
+  Home,
+  Jobs,
+  JobOffers,
+  Projects,
+  Tasks,
+  Users
+}